chore: do not add source when testing CombinedHandler
Since the source is always going to be different in the runner, it is better to not add the source since it doesn't add much to the test itself. The basic function of whether the handler is being initialized with different configs is taken care of by the log levels being different.

Signed-off-by: ksdfg <ksdfg123@gmail.com>